Version 1.5

+++ app.py
@@ -1,15 +1,17 @@
 import json
+import os
 import secrets
 from datetime import datetime
 
 import flask
 from flask import Flask, redirect, render_template, url_for
 from flask_login import login_required, current_user, login_user, logout_user
-
+from twilio.base.exceptions import TwilioRestException
+from twilio.rest import Client
 from data_handler.pdf_handler import pdf_handler
-from instance.models import db, login, UserModel, ItemsModel, LocationsModel, RequestsModel
+from instance.models import db, login, UserModel, ItemsModel, LocationsModel, RequestsModel, TrackersModel
 
-verification_codes = {}
+notif_numbers = []
 app = Flask(__name__)
 app.secret_key = 'Smkc01002dazzxqn3'
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
@@ -19,6 +21,7 @@ with app.app_context():
     db.create_all()
 login.init_app(app)
 login.login_view = 'login'
+client = Client(os.getenv('TWILIO_ID'), os.getenv('TWILIO_TOKEN'))
 
 
 @app.route('/', methods=['GET', 'POST'])
@@ -57,10 +60,8 @@ def requests():
                 db.session.add(item)
         db.session.commit()
         # send out the mobile notif
-        # statement = select(mn.mobile_notif)
-        # notif_numbers = session.exec(statement).fetchall()
-        # mn.send_notif(config['TWILIO_ID'], config['TWILIO_TOKEN'], config['TWILIO_MSG_SID'], notif_numbers)
-        return redirect('/home')
+        send_notif()
+        return redirect('/')
     elif flask.request.method == 'GET':
         locations = LocationsModel.query.all()
         transfer_requests = RequestsModel.query.all()
@@ -116,15 +117,21 @@ def items():
     mode = flask.request.args.get('mode')
     if mode == 'add' and current_user.user_type == 'admin':
         name = flask.request.args.get('name')
+        stock_item = ItemsModel.query.filter_by(name=name).first()
+        if stock_item:
+            return 'An item with that name already exists.'
         stock_item = ItemsModel(name=name)
         db.session.add(stock_item)
         db.session.commit()
+        return redirect('/items')
     elif mode == 'delete' and current_user.user_type == 'admin':
         item_ids = flask.request.args.get('ids').removesuffix(":").split(':')
         for item_id in item_ids:
             item = ItemsModel.query.filter_by(item_id=item_id).first()
-            db.session.delete(item)
+            if item:
+                db.session.delete(item)
         db.session.commit()
+        return redirect('/items')
     stock_items = ItemsModel.query.all()
     return render_template('authenticated/basic/items/stock_items.html', stock_items=stock_items)
 
@@ -133,7 +140,13 @@ def items():
 @login_required
 def item(item_id: int):
     if flask.request.method == 'POST':
-        pass
+        name = flask.request.json['name']
+        item = ItemsModel.query.filter_by(item_id=item_id).first()
+        if item:
+            item.name = name
+            db.session.add(item)
+        db.session.commit()
+        return redirect(f'/item/{item_id}')
     elif flask.request.method == 'GET':
         item = ItemsModel.query.filter_by(item_id=item_id).first()
         if item:
@@ -143,7 +156,7 @@ def item(item_id: int):
                 request = RequestsModel.query.filter_by(request_id=request_id).first()
                 if request:
                     request_history.append(request)
-            return render_template('authenticated/basic/items/stock_item.html', item=item, history=request_history)
+            return render_template('authenticated/basic/items/stock_item.html', item=item, history=request_history.__reversed__())
     return render_template('public/404.html')
 
 
@@ -154,31 +167,43 @@ def locations():
         mode = flask.request.args.get('mode')
         if mode == 'add' and current_user.user_type == 'admin':
             name = flask.request.json['name']
+            stock_location = LocationsModel.query.filter_by(name=name).first()
+            if stock_location:
+                return 'That location already exists.'
             stock_location = LocationsModel(name=name)
             db.session.add(stock_location)
             db.session.commit()
             return redirect('/locations')
         elif mode == 'delete' and current_user.user_type == 'admin':
-            location_ids = flask.request.args.get('ids').removesuffix(":").split(':')
+            location_ids = flask.request.args.get('ids').removesuffix(':').split(':')
+            print(location_ids)
             for location_id in location_ids:
                 location = LocationsModel.query.filter_by(location_id=location_id).first()
-                db.session.delete(location)
+                if location:
+                    db.session.delete(location)
             db.session.commit()
+            return redirect('/locations')
     elif flask.request.method == 'GET':
         stock_locations = LocationsModel.query.all()
-        return render_template('authenticated/basic/locations/stock-locations.html', stock_locations=stock_locations)
+        return render_template('authenticated/basic/locations/stock_locations.html', stock_locations=stock_locations)
 
 
 @app.route('/location/<location_id>', methods=['POST', 'GET'])
 @login_required
 def location(location_id: int):
     if flask.request.method == 'POST' and current_user.user_type == 'admin':
-        pass  # delete the location?
+        name = flask.request.json['name']
+        location = LocationsModel.query.filter_by(location_id=location_id).first()
+        if location:
+            location.name = name
+            db.session.add(location)
+        db.session.commit()
+        return redirect(f'/location/{location_id}')
     elif flask.request.method == 'GET':
         location = LocationsModel.query.filter_by(location_id=location_id).first()
         if location:
             transfers = RequestsModel.query.filter_by(location_id=location.location_id)
-            return render_template('authenticated/basic/locations/view-location.html', location=location, history=transfers)
+            return render_template('authenticated/basic/locations/stock_location.html', location=location, history=transfers)
     return render_template('public/404.html')
 
 
@@ -238,15 +263,17 @@ def database():
         pdfh = pdf_handler('data_handler/stock_items_summary.pdf')
         items = pdfh.import_venue_stock_items()
         for item in items:
-            if not item.startswith('Delete'):
-                item = ItemsModel(name=item)
+            name = item.strip()
+            item = ItemsModel.query.filter_by(name=name).first()
+            if not name.startswith('Delete') and not item:
+                item = ItemsModel(name=name)
                 db.session.add(item)
         try:
             db.session.commit()
         except Exception as e:
             print(e)
             db.session.rollback()
-        return redirect(url_for('admin'))
+        return redirect('/items')
     if mode == 'init-locs':
         with open("data_handler/locations_data.txt") as file:
             lines = file.readlines()
@@ -258,7 +285,7 @@ def database():
                 db.session.commit()
             except:
                 db.session.rollback()
-        return redirect(url_for('admin'))
+        return redirect('/locations')
     if mode == 'init-dev-users':
         print('init dev')
         user = UserModel(employee_id=1061, name='Bryce Cotton', user_type='admin', active=1)
@@ -266,6 +293,101 @@ def database():
         db.session.add(user)
         db.session.commit()
         return redirect('/login')
+    if mode == 'backup-users':
+        users = UserModel.query.filter_by(active=True)
+        with open('users.txt', 'a+') as file:
+            for user in users:
+                jsonstr = user.to_json()
+                print(jsonstr)
+                file.write(f'{jsonstr}\n')
+        return redirect('/login')
+    if mode == 'load-users':
+        with open('users.txt', 'r+') as file:
+            users = [x.strip() for x in file.readlines()]
+            for user in users:
+                user = json.loads(user)
+                employee_id = user['employee_id']
+                name = user['name']
+                hash = user['hash']
+                user_type = user['user_type']
+                db_user = UserModel.query.filter_by(employee_id=employee_id).first()
+                if db_user:
+                    print('USER EXISTS ALREADY')
+                else:
+                    db_user = UserModel(employee_id=employee_id, name=name, active=True, user_type=user_type)
+                    db_user.set_password(password=hash, use_hash=False)
+                    db.session.add(db_user)
+            db.session.commit()
+        return redirect('/login')
+
+
+
+@app.route('/mobile_notif/<employee_id>', methods=['POST'])
+@login_required
+def mobile_notif(employee_id: int):
+    if flask.request.method == 'POST':
+        user = UserModel.query.filter_by(employee_id=employee_id).first()
+        option = flask.request.json['option']
+        if user:
+            if option == 'enable':
+                user.enable_mobile_notifications = True
+            elif option == 'disable':
+                user.enable_mobile_notifications = False
+            number = user.phone
+            if number:
+                send_notif(number=number, msg=f'You have {option}d mobile notifications!')
+            db.session.add(user)
+        db.session.commit()
+        return redirect(f'/profile/{employee_id}')
+    elif flask.request.method == 'GET':
+        return render_template('public/404.html')
+
+
+@app.route('/trackers', methods=['POST', 'GET'])
+@login_required
+def trackers():
+    if current_user.user_type == 'admin':
+        if flask.request.method == 'POST':
+            items = flask.request.json['items'].removesuffix(':').split(':')
+            for item in items:
+                tracker = TrackersModel.query.filter_by(tracker_id=item).first()
+                db.session.delete(tracker)
+            db.session.commit()
+            return redirect('/trackers')
+        elif flask.request.method == 'GET':
+            trackers = TrackersModel.query.all()
+            return render_template('authenticated/basic/tools/expiry_date_tracker/trackers.html', trackers=trackers)
+    else:
+        if flask.request.method == 'POST':
+            return 'this route does not exist.'
+        elif flask.request.method == 'GET':
+            return render_template('public/404.html')
+
+
+@app.route('/create_tracker', methods=['POST', 'GET'])
+@login_required
+def create_tracker():
+    if current_user.user_type == 'admin':
+        print('test')
+        if flask.request.method == 'POST':
+            name = flask.request.json['name']
+            date_received = flask.request.json['date_received']
+            expire_date = flask.request.json['expire_date']
+            quantity = flask.request.json['quantity']
+            employee_id = current_user.get_id()
+            tracker = TrackersModel(employee_id=employee_id, date_received=date_received, expiry_date=expire_date, quantity=quantity, name=name)
+            db.session.add(tracker)
+            db.session.commit()
+            return redirect('/trackers')
+        elif flask.request.method == 'GET':
+            stock_items = ItemsModel.query.all()
+            return render_template('authenticated/basic/tools/expiry_date_tracker/create_tracker.html', stock_items=stock_items)
+    else:
+        print('test 2')
+        if flask.request.method == 'POST':
+            return 'this route does not exist.'
+        elif flask.request.method == 'GET':
+            return render_template('public/404.html')
 
 
 @app.route('/admin', methods=['POST', 'GET'])
@@ -302,14 +424,22 @@ def login():
 
     if flask.request.method == 'POST':
         employee_id = flask.request.json['employee_id']
-        user = UserModel.query.filter_by(employee_id=int(employee_id)).first()
-        if user and user.active and user.check_password(flask.request.json['password']):
-            login_user(user)
-            return redirect(f'/profile/{employee_id}')
+        if employee_id:
+            user = UserModel.query.filter_by(employee_id=int(employee_id)).first()
+            if user and user.active and user.check_password(flask.request.json['password']):
+                login_user(user)
+                return redirect(f'/profile/{employee_id}')
+            else:
+                return 'Invalid User or Credentials.'  # make the page display a note for the user, so they know tiw as a fail
         else:
-            return 'Invalid User or Credentials.'  # make the page display a note for the user, so they know tiw as a fail
+            return 'Username was left blank.'
 
     return render_template('public/login.html')
+
+
+@app.route('/test', methods=['POST', 'GET'])
+def test():
+    return render_template('test.html')
 
 
 @app.route('/logout')
@@ -340,6 +470,27 @@ def get_location_name(location_id):
         return ''
 
 
+def send_notif(number: str = '', msg: str = ''):
+    if msg == '':
+        msg = 'Transfer Request Received!'
+    if number == '':
+        users = UserModel.query.filter_by(enable_mobile_notifications=True).all()
+        print(users)
+        for user in users:
+            phone = user.phone
+            print(f'phone = {phone}')
+            if phone:
+                try:
+                    client.messages.create(messaging_service_sid=os.getenv('TWILIO_MSG_ID'), from_='+12044002185', body=f"TO: {number}\n\n{msg}\n\nFROM: Bryce's Transfer Service",
+                                           to=f'+1{phone}')
+                except TwilioRestException as e:
+                    print(f'Failed to send text to {number}\n{e}')
+    else:
+        try:
+            client.messages.create(messaging_service_sid=os.getenv('TWILIO_MSG_ID'), from_='+12044002185', body=f"TO: +1{number}\n\n{msg}\n\nFROM: Bryce's Transfer Service", to=f'+1{number}')
+        except TwilioRestException as e:
+            print(f'Failed to send text to {number}\n{e}')
+
+
 if __name__ == '__main__':
-    # app.run(host='0.0.0.0', port=8000)
     app.run(host='0.0.0.0', port=5000, debug=True)
